remove remaining occurrence of useLogout, propagate location state from /logout to /welcome
