+++ src/modals/DeactivateAccountModal.tsx
@@ -7,7 +7,6 @@ import { useTranslation } from 'react-i18next';
 import { useNavigate } from 'react-router-dom';
 import { useUserType } from '../hooks/useApollo';
 import DisableableButton from '../components/DisablebleButton';
-import useLogout from '../hooks/useLogout';
 
 // corresponding dissolve reason ids in translation file
 // for now just loop through 0-5 and 0-6 (+1 in loop)
@@ -25,7 +24,6 @@ const DeactivateAccountModal: React.FC<Props> = ({ isOpen, onCloseModal }) => {
     const { space } = useTheme();
     const navigate = useNavigate();
     const { trackEvent } = useMatomo();
-    const logout = useLogout();
     const { t } = useTranslation();
 
     const userType = useUserType();
@@ -76,15 +74,14 @@ const DeactivateAccountModal: React.FC<Props> = ({ isOpen, onCloseModal }) => {
                     name: 'Account deaktivieren',
                     documentTitle: 'Deactivate',
                 });
-                logout();
-                navigate('/welcome', { state: { deactivated: true } });
+                navigate('/logout', { state: { deactivated: true } });
             } else {
                 showError();
             }
         } catch (e) {
             showError();
         }
-    }, [reason, deactivateAccount, isOther, t, userType, other, onCloseModal, trackEvent, logout, navigate, toast]);
+    }, [reason, deactivateAccount, isOther, t, userType, other, onCloseModal, trackEvent, navigate, toast]);
 
     const isValidInput = useMemo(() => {
         if (!reason) return false;
